# cooperation/src/main/resources/static/doc/openapi_v2.yaml
openapi: 3.0.1
info:
  title: Etablerad samverkan i vården
  description: "![Informationsmodell - Cooperation](images/cooperation.png)\r\n\r\nInformation om tjänster, producenter och konsumenter baserad på tjänsteadresseringskatalogen i både Ineras nationella tjänsteplattform och regionala tjänsteplattformar. Eventuella framtida konfigurationsändringar syns inte i api:et, enbart hur det ser ut i dagsläget. Konfiguration i en tjänsteplattform kan ibland ske på en viss nivå i HSA-trädet, och avser alla noder därunder. Detta speglas rakt av i api:et och HSA-trädet evalueras inte här. De tekniska id-begreppen på alla objekt genereras om en gång per dygn och garanteras således inte ha samma värde två dagar i rad. Applikationer bör skrivas så att de är oberoende av de tekniska nycklarna. Varje resurs har sin egen nummerserie för de tekniska id-begreppen. En api-nyckel kan behövas för att nå api:et."
  version: 2.0.0
servers:
  - url: /coop/api/v2
paths:
  /connectionPoints:
    get:
      summary: Connection Point - TP instans
      description: |
        Alla instanser av tjänsteplattformar som har exporterat information till detta api. Idag sker exporten en gång per dygn. En instans identifieras verksamhetsmässigt av kombinationen mellan en plattform och vilken staging-miljö som avses.
      parameters:
        - name: platform
          in: query
          description: Filtrera på plattform, tex NTJP
          required: false
          schema:
            type: string
        - name: environment
          in: query
          description: Filtrerar på miljö tex PRODUCTION
          required: false
          schema:
            type: string
        - name: serviceConsumerId
          in: query
          description: Filtrerar bara ConnectionPoints som innehåller någon samverkan med en specifik tjänstekonsument. Ange den tekniska nyckeln för serviceConsumer tex 538.
          required: false
          schema:
            type: integer
            format: int64
        - name: logicalAddressId
          in: query
          description: Filtrerar bara ConnectionPoints som innehåller någon samverkan med en specifik logisk adress. Ange den tekniska nyckeln för LogicalAddress tex 42.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceContractId
          in: query
          description: Filtrerar bara ConnectionPoints som innehåller någon samverkan med en specifik tjänstekontrakt. Ange den tekniska nyckeln för ServiceContract tex 123.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceProducerId
          in: query
          description: Filtrerar bara ConnectionPoints som innehåller någon samverkan med en specifik tjänsteproducent. Sökningen utförs genom att titta på tjänstproducenter som publicerar en tjänst på en viss logisk adress samtidigt som någon tjänstekonsument är konfigurerad att anropa samma logiska adress.  Ange den tekniska nyckeln för ServiceProducer tex 538.
          required: false
          schema:
            type: integer
            format: int64
      tags:
        - ConnectionPoint
      responses:
        '200':
          description: En lista av tjänsteplattformsinstanser
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectionPoint'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConnectionPoint'
                xml:
                  wrapped: true
                  name: connectionPoints
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /connectionPoints/{id}:
    get:
      summary: Connection Point - TP instans
      description: |
        En specifik instans av tjänsteplattformen.
      parameters:
        - $ref: '#/components/parameters/TechnicalKeyParam'
      tags:
        - ConnectionPoint
      responses:
        '200':
          description: En ConnectionPoint
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionPoint'
            application/xml:
              schema:
                $ref: '#/components/schemas/ConnectionPoint'
        '404':
          description: ConnectionPoint not found
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
...
  /cooperations:
    get:
      summary: Cooperation - samverkan
      description: |
        Alla samverkan. En tjänstekonsument är konfigurerad att samverka via en logisk adress avseende ett tjänstekontrakt. Objektet i sig innehåller inga attribut men genom att använda include-parametern hängs det på information från 1-4 relaterade objekt som beskriver samverkan. OBS prova inte denna resurs på denna sida utan att ange minst en filtrerings-parameter. Det är ganska stor träffbild om man får alla förekomster tillbaka och de skall parsas i dokumentverktyget, tar lång tid.  
      parameters:
        - name: connectionPointId
          in: query
          description: Filtrerar bara samverkan som är konfigurerad i den aktulla instansen av tjänsteplattformen. Ange den tekniska nyckeln för connectionPoint tex 376.
          required: false
          schema:
            type: integer
            format: int64
        - name: logicalAddressId
          in: query
          description: Filtrerar bara samverkan som avser en specifik logisk adress. Ange den tekniska nyckeln för LogicalAddress tex 42.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceContractId
          in: query
          description: Filtrerar bara samverkan som avser ett specifik tjänstekontrakt. Ange den tekniska nyckeln för ServiceContract tex 123.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceConsumerId
          in: query
          description: Filtrerar bara samverkan avseende en specifik tjänstekonsument. Ange den tekniska nyckeln för ServiceConsumer tex 538.
          required: false
          schema:
            type: integer
            format: int64
        - name: serviceDomainId
          in: query
          description: Filtrerar bara samverkan avseende alla kontrakt som ingår i en viss tjänstedomän. Ange den tekniska nyckeln för ServiceDomain tex 825.
          required: false
          schema:
            type: integer
            format: int64
        - name: include
          in: query
          description: 'Styr hur mycket information som skall skickas tillbaka i svaret. Ange en eller flera av följande objekt kommaseparerade: connectionPoint,logicalAddress,serviceContract,serviceConsumer'
          required: false
          schema:
            type: string
            default: connectionPoint,logicalAddress,serviceContract,serviceConsumer
      tags:
        - Cooperation
      responses:
        '200':
          description: En lista av samverkan
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cooperation'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cooperation'
                xml:
                  wrapped: true
                  name: cooperations
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
  /cooperations/{id}:
    get:
      summary: Cooperation - samverkan
      description: |
        En specifik samverkan. Alla 4 relaterade objekt är alltid inkluderade.
      parameters:
        - name: id
          in: path
          description: Teknisk nyckel
          required: true
          schema:
            type: number
      tags:
        - Cooperation
      responses:
        '200':
          description: En samverkan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cooperation'
            application/xml:
              schema:
                $ref: '#/components/schemas/Cooperation'
        '404':
          description: Cooperation not found
        default:
          $ref: '#/components/responses/UnexpectedErrorResponse'
...
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic
  responses:
    UnexpectedErrorResponse:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
        application/xml:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    TechnicalKeyParam:
      name: id
      in: path
      description: Teknisk nyckel
      required: true
      schema:
        type: number
  schemas:
    ConnectionPoint:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        platform:
          type: string
          description: Plattform tex NTJP.
        environment:
          type: string
          description: Miljö tex PRODUCTION.
        snapshotTime:
          type: string
          description: Den tidpunkt som informationen lästes av från tjänsteplattformen.
      xml:
        name: connectionPoint
    ServiceConsumer:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        hsaId:
          type: string
          description: hsa-id.
        description:
          type: string
          description: Beskrivning.
      xml:
        name: serviceConsumer
    Cooperation:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        serviceConsumer:
          $ref: '#/components/schemas/ServiceConsumer'
        logicalAdress:
          $ref: '#/components/schemas/LogicalAddress'
        serviceContract:
          $ref: '#/components/schemas/ServiceContract'
        connectionPoint:
          $ref: '#/components/schemas/ConnectionPoint'
      xml:
        name: cooperation
    LogicalAddress:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        logicalAddress:
          type: string
          description: logisk adress.
        description:
          type: string
          description: Beskrivning.
      xml:
        name: logicalAddress
    ServiceContract:
      type: object
      properties:
        id:
          type: number
          description: Unik teknisk nyckel
        name:
          type: string
          description: Namn.
        namespace:
          type: string
          description: Namnrymd.
        major:
          type: number
          description: Major-version exvis 1 eller 2.
        minor:
          type: number
          description: Minor-version exvis 0 eller 1.
      xml:
        name: serviceContract
...
    Error:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        fields:
          type: string
      xml:
        name: error